requirements.txt
opencv-python-headless==4.11.*
opencv-contrib-python==4.11.*
numpy==1.26.*
scikit-image==0.23.*
pillow==10.*
kraken==3.*
pyyaml==6.*
tqdm==4.*
pandas==2.*
# Optional
wandb
# Development
pytest==7.*
black==23.*
flake8==6.*
